MSVC use asm x86-64 code instead of C++ implementation
The goal is to have performance close to GCC build (MSVC is near 10% slower than GCC with current C++ code). Disadvantage is that this asm code is tuned for Intel Core 2 processor, with them the performance is now slightly better (1-2%) than GCC ! With other processors, MSVC and GCC builds must be compared.